search-areaをコンポーネント化
